CA-184556: call VBD actions in parallel to speed up VM lifecycle operations
such as vm-start, vm-shutdown, vm-restore, vm-poweroff, vm-reboot.

In this patch, a new composable Parallel atom in xenopsd is created. Any
atoms inside Parallel will be executed concurrently in a new independent
pool of worker threads specific for the Parallel atoms. A new pool of
parallel worker threads is used to avoid a deadlock where an operation
needs to wait for a Parallel atom to run but the Parallel atom cannot run
because all the default worker thread slots are being used by other
operations also waiting for Parallel atoms to run.

The Parallel atom is used to execute in parallel all the VBD-related atoms
that are otherwise taking a long time to execute as a sequence during the
VM lifecycle operations, which are:

    VBD.epoch_begin
    VBD.plug
    VBD.unplug
    VBD.epoch_end

The SM layer supports the parallel execution of these operations, since
this is already the case during the parallel execution of VM lifecycle
operations. We strived to achieve safety in xenopsd by limiting the
concurrency to only inside the Parallel atoms, and by carefully looking
the code inside the xc/VBD.{plug,unplug,epoch_begin,epoch_end} which are
related to accessing potentially shared datastructures.

Signed-off-by: Marcus Granado marcus.granado@citrix.com