CA-184556: deactivate the datapath devices in parallel
During VM.save in vm-migrate, code specific to xc deactives each
datapath device. A VM can have many such devices, taking many
minutes to deactivate all of them.

This patch detaches these devices in parallel. Unfortunately, it's
not possible to reuse the Parallel atom because this code is specific
to xc and not in the generic xenopsd control path code. However,
it's possible to simulate Parallel using Thread.thread_iter on small
chunks of vbds.

Signed-off-by: Marcus Granado <marcus.granado@citrix.com>