CA-195652: Remove redundant refresh_vm calls
Now that Events_from_xenops.with_suppressed is the only way to be sure
that a refresh_vm is called with events enabled (since after it
completes it holds the lock until refresh_vm has completed), then we can
pass a no-op function to with_suppressed to cause the syncing we
require in maybe_cleanup_vm.

This patch also removes other calls to refresh_vm that are now redundant
(because they were called after an Events_from_xenopsd.with_suppressed).

There are a few other places where refresh_vm which we should consider
revising since it is not guaranteed to have the desired effect if
another thread has suppressed events.

The one place where calling refresh_vm is valid is in the
pool_migrate_complete call which is executed on the destination host of
a migration.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>