CA-197914: Avoid deadlock by ignoring suppression in set_resident_on
By waiting for event suppression to finish in set_resident_on,
localhost migration was deadlocking because pool_migrate_complete
was called with events suppressed on the sender, and this was
in turn calling set_resident_on.

This fix simply ignores event suppression during set_resident_on,
which is safe because we only call set_resident_on when the VM
is just starting, and so we should be protected by the VM's
current_operations.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>