CA-201099: Read PIDs from pidfiles, rather than calling pidof
RRDD used to call pidof to find the processes it was monitoring.
This imposed a slight extra load in dom0, and if a particular
process was not running pidof would fail and RRDD would log the
failure, causing lots of spurious logging.

Now we get the pids from pidfiles, and handle a missing pidfile
silently.

Signed-off-by: Euan Harris <euan.harris@citrix.com>