CA-202385: SR assert_operation_valid: only run relevant checks
In Xapi_sr_operations.assert_operation_valid, we know which operation we're
trying to perform, so don't bother checking errors that would occur for other
operations. That's a waste of energy.

Since some of the checks involve expensive operations, this can improve
performance. Measurements confirm that the duration of a vdi-create operation
when there are 16,000 VDIs is reduced from 4.8 to 3.3 seconds, i.e. a 30%
improvement.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>