CA-200924: Avoid premature cache invalidation
The 2 caches (xapi and xenopsd) should be populated whenever there
is metadata in xenopsd. This fix populates it at start-of-day
by asking xenopsd which VMs it knows about, and prevents premature
removal of the caches. Previously the 'maybe_cleanup_vm' function
was removing the caches whether or not it managed to cleanup the
VM from xenopsd.

The caches are now removed in only one place - when the VM metadata
is removed from xenopsd.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>