CA-200924: Suppress xapi events a bit longer than xenopsd
Any changes to the xapi metadata during the xenopsd resync that's
kicked off following a suppression will cause the xapi events thread
to fire. If the xapi events thread sees the 'currently_attached' field
before it's been synced by the xenopsd events, it will cause an
unnecessary exception and a delay of 15 seconds. So we suppress the
xapi events thread until after the xenopsd events have been synced.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>