CA-203414: Upgrade CPU features in pool_migrate_complete after SXM
The last_boot_CPU_flags field of a VM that is migrated to a newer-version host
in another pool (SXM) was not updated in the xapi DB. This would not make the
migration itself fail, because the receiving xenopsd upgrades the featureset
itself before restoring the domain. However, after the migration, xapi would
overwrite the featureset in xenopsd's VM metadata with the one from the xapi
DB, which was still out-of-date. For this reason, a following reboot may fail.

This patch ensures that the last_boot_CPU_flags field is updated immediately
after a cross-pool migration finishes, in the pool_migration_complete call.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>