CP-16558: assert_can_set_has_vendor_device: don't refer to *_paths_optimized
The check in assert_can_set_has_vendor_device of whether any network paths or
storage paths are optimized is a guess at best. Remembering whether the PV
drivers were present when the VM was last shut down is not a fool-proof way of
determining whether they will be present when the VM next boots (e.g. you could
have attached the VDI elsewhere.) Whatever kind of check we do will be
imperfect.

Moreover, this check is not protecting against any bad situations. The PV
drivers will not unplug the emulated devices, so the VM will remain functional.

This check adds complication to the logic that is used to show the user whether
PV drivers are not installed -– see CA-204963.

Hence, let's remove the check.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>