CA-203169: don't return entire device tree from VM.stat
Currently, VM.stat returns the contents of a number of a xenstore subtrees
relating to the guest.

Since xapi no longer relies on any of the data from the 'devices' subtree,
xenopsd no longer needs to include this in the VM.stat response.

This gives rise to a significant performance benefit, as the amount of xenstore
traffic is significantly reduced, especially for VMs with several VBDs or VIFs.

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>