CA-205589: Preserve has_vendor_device on DB upgrade
We already had a DB upgrade rule to set VM.has_vendor_device to false
on upgrade from a pre-Dundee version of the DB schema.

Now instead of "pre-Dundee", the rule is applied if the schema version
is older than the first tech-preview in which the has_vendor_device
field controlled the presence of the device. (In the previous schema
version and tech preview, the field was present but had no effect.)

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>