CA-203169: don't fail DB upgrade from Tech Previews that drop fields
Since the previous commit removed some fields from the datamodel, upgrade from
the Dundee Tech Preview, which included those fields, will fail. This patch
makes the upgrade proceed and silently drop these fields.

In general, Tech Preview releases may contain fields that are subsequently
dropped. Hence a list of such versions is added to the datamodel. This list is
consulted by the unmarshalling code. (This approach is preferred to hard-coding
the names of the dropped fields in the unmarshalling code.)

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>