CA-206753: Move task cancellation to message_forwarding
We need task cancellation of vm lifecycle operation in
message_forwarding in order to aquire lock on VBDs and VIFs
for hard_reboot and hard_shutdown operations.

And exception raised during shutdown and reboot must
wait for the event thread to sync on slave before returning
in order to update the state of VM on slave.

Signed-off-by: Sharad Yadav <sharad.yadav@citrix.com>