CA-16625 powercycle test back to old behavior when no UUID given
The powercycle test is run by default and needs a VM for this. It has in
the past relied on finding a suitable VM by name (but given up silently,
I suspect, when it didn't find one). We want to move to a scheme where
any VM used is specified by a UUID template provided by -template-uuid.
When no such UUID is provided, powercycle now falls back to the old
behavior.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>