CA-209511: Don't remove private data path for devices during localhost migration
There can be multiple domains for a single VM during a localhost migration, and
the private data path is indexed by UUID, not domid. When unplugging the
devices for the old domain during a localhost migration, we must not remove the
private data keys, because they are needed by the new domain.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>