CA-199734: best-effort DP destroy and other cleanups for mirror/copy task failure
... previously we didn't do this which caused issue like CA-199734.

A bit of code re-organization and refactoring is done as part of this. Mainly,
we use reference cells to record configuration variables. These ref cells are
gradually filled up with actual configurations along the process, so that the
initialization status of these cells also reflects the overall progress, i.e.
which steps have already been done and which haven't. With this information,
the error handling logic can decide what to do or undo in case of failures. The
benefits are, there is no need for multi layers of exceptions handling, also a
single entry point for a unified cleanup function.

Signed-off-by: Zheng Li <dev@zheng.li>