CA-210925: front-to-back device links under /xenops
The problem fixed by this commit:
When shutting down a guest, we were deriving the list of devices
to be cleaned up from the values of the guest domain's nodes at
/local/domain/N/device/.../backend in xenstore; these may have
been updated by the guest, in which case the shutdown could leave
the guest domain in a bad state.

The fix:
Now when we add a device, as well as writing to
/local/domain/N/device/.../backend
we also write the same to
/xenops/domain/N/device/.../backend
and then refer to this reliable data where appropriate:
the whole /xenops tree is accessible by dom0 only.

(This is the final commit of a series of three.)
Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>