CP-17469: Xapi_pci: Treat all known PCI classes as managed
This means that the VM.attached_PCIs will be updated when the VM is
started and shutdown and this will inform the VM.allowed_operations.

For example, VMs with PCI-passthrough (using the other-config:pci key)
will have the same operation restrictions as a VM using GPU-passthrough
(previously the only class of managed PCI device). Presently the
operations restricted based on attached PCI devices are:

* suspend
* checkpoint
* pool_migrate
* migrate_send

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>