CA-105021: Set Console.location to empty if Host.address is empty
The usual form for the console URL is:

    https://<host-ip-address>/console?ref=<console-ref>

However, when the host doesn't have an IP address because, for example,
the host management network has been disabled, then the console URL
becomes

    https:///console?ref=<console-ref>

which doesn't make sense. It doesn't make sense, either, to maintain the
IP address that the host _used_ to have because this IP might be in use
somewhere else if, for example, there was a DHCP server running on the
network.

This patch sets the location URL to the empty string if the host doesn't
have an address on the network.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>