Revert "Wait for the same hotplug path as upstream scripts, in all cases"
This reverts commit 624a29f407b2d63312352bc387061b1e09238b16, so that xenopsd
(as in the past) waits for a key from the "private" xenstore tree to be removed
by the hotplug script, before removing the rest of the private keys.

This way, the private keys are still available to the hotplug script for any
cleanup actions. By this time, the backend keys in /local/domain have been
removed already (which is what triggered the hotplug script to be called in the
first place), so that any information in there can no longer be used.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>