CP-18173 add check to vm lifecycle: no migration when using nested virt
The check is a little more exhaustive: Migration, suspension, and
checkpointing must be blocked if either nested virtualisation is active
or platform:nomigrate is true. This can be overridden with --force.
The commit contains unit tests.

This commit will be updated to use a library function that interprets
boolean values from platform flags.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>