CP-18172 fix: catch exception when parsing last booted record
To obtain the last-booted record, this commit now uses
Helpers.get_boot_record_of_record instead of Helpers.parse_boot_record
This fixes the problem that VM.resume was failing because of an
exception that was escaping from parse_boot_record when no last boot
record exists but that is caught by get_boot_record_of_record. The
semantics of get_boot_record_of_record seems dubious as it in that case
returns the current record.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>