CP-18173 consider actual platform flags, not those from last-booted record
It turned out that xapi doesn't have access to the last-booted record
and that existing code for accessing it is actually reporting the
current record. This leads still to the correct result unless a user
manipulates the nested-virt value in the platform field after booting.
The proper last boot record is maintained by xenopsd.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>