CA-216726: mechanism to persistent xenopsd metadata to disk
If we need to have VMs started before xapi (for example, if we've got HA
enabled using an SR that requires a VM up to attach it), then we need to
maintain a copy of the xenopsd metadata that we can push to it on system
start. This needs to be kept up-to-date as xapi's metadata changes or
things will become inconsistent.

Rather than implementing another database watcher, this commit adds some
code to the existing thread that updates xenopsd metadata when xapi's db
changes. It's conditional on a kv pair appearing in the VM's other-config
which should be of the form:

    other-config:persist_xenopsd_md=<filename>

The filename is relative to /var/lib/xcp/xenopsd_md/

A caveat to this is that this file is only kept up to date on the host
on which the VM is running. This is almost certainly sufficient for the
use-case described above.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>