CA-218221: make sure SXM destination VDI has the same size as the source
When SXM copies a VDI across, it first tries to find a pair of VDIs with
identical contents on both sender and receiver sides as base copies. And the
content of the pair should also be close enough to the content of source VDI.

This way, we can simply make a clone (VHD child) of the base copy on the
reciver side, which will then become the destination VDI. We then only need to
copy into it the small differences between the source VDI and the base copy on the
sender side.

In some cases, despite the similarity, base copies might have a size different
from the source VDI's. This is usually true when there was a resize operation
in the history of VHD chain. In these cases, the destination VDI (as the result
of direct cloning the base copy on the reciever side) will have a size
initially different from the source VDI, therefore a resize operation is
necessary to make sure the destination VDIs having the same side as the source.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>