CA-212079: Allow XenMotion and AD during RPU from pre-Dundee
The XenMotion and AD feature recently had restrictions put in place, so that
they can be switched on and off by feature flags. A feature is considered to be
enabled only if all hosts in the pool have it enabled. During a rolling pool
upgrade (RPU) from a pool that does not have the new feature flags, the absence
of the flags was treated as a feature restriction, such that the pool as a
whole had XenMotion and AD restricted. Luckily the restrictions were only
enforced on hosts that were already upgraded, so that VMs could still migrate
off hosts that still need to be upgraded.

This patch introduces a list of features (currently just XenMotion and AD) that
are considered to be enabled if the feature flag is absent, to avoid the issues
described above.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>