CA-217842: Replace instances of vm_lacks_feature_x with vm_lacks_feature
Rationale: This leads to less confusion when, for example, attempting to
migrate a VM leads to an error that the VM does not support suspend. The
lacking feature is currently always better defined by context than by the
error text itself, so let's make the error generic and rely on the
context to signify which feature is lacking.

Also, given that the shutdown/reboot features might be advertised by either
feature_poweroff,feature_reboot or feature_shutdown, this makes the
error raised less wrong.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>