CP-18476: Let setup-vif-rules only remove its own flows
When cleaning up, the setup-vif-rules script deletes all flows for a given
in_port. This is a little rude, because there may exist flows that match on the
same in_port for different purposes (e.g. from setup-pvs-proxy-rules). Use a
cookie as a tag to identify flows that were added by setup-vif-rules itself.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>