CA-219763: If the PV drivers have been detected, raise an update
Since this is an update to a VIF or VBD, the usual logic only causes
an update for the relevant device. Thus xapi didn't know to look at
the VM again, and hence might not update the pv_drivers_detected
field. This commit adds an explicit update to the VM if the
pv_drivers_detected field changes from false to true.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>