CP-18294 rewrite unit test for mobility testing
This series of commits changes the way VM mobility is decided: the
relevant value are now picked up from VM_metrics and no longer from
platform flags. This must be reflected in the tests. The test suite now
populates the relevant fields in VM_metrics and tests the critical
VM operations.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>