CA-206623: close the input end of the pipe when data feeding is complete
Without the patch, the feeder side won't close the pipe's entrance when all the
data has been streamed in. It relys on the consumer to do that eventually. So
in case when the feeder encounters a premature EOF in its own input stream (say
the link breaks), that EOF effect won't be passed through the pipe as a result.
So that the consumer side will be waiting forever for the data yet to come, and
the whole import process will stall.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>