Set the correct vif device id or fail
When updating the MTU of a `VIF` it can happen that the VM `domid` is still out of sync and consequently the computed device id is wrong.
We can use the correct device value passing it via `xenopsd` as a field in `VIF.stat`.

See also https://github.com/xapi-project/xcp-idl/pull/131 and https://github.com/xapi-project/xenopsd/pull/288

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>