CA-220466: Guest reads/deletes in dom0 via grub config
Bug: a guest can write a grub config that will cause an arbitrary dom0
file to be made available to the guest as a ramdisk, and then deleted.

Fix: use output-format=simple for pygrub (instead of the sexp-based
output format).

This change also expects the same line-based format from eliloader, so
it depends on a corresponding change to eliloader, which needs to emit
its output in the line-based "simple" format.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>