When asserting a VM can be migrated, check first the target host version.
If the version check fails, we should not show the same error message for intra-
and cross-pool migration, because it does not apply to the latter case.
Therefore, add a new api_error, remove the
Helpers.assert_host_versions_not_decreasing wrapper (not much of a shortcut
anyway) and display the appropriate message each time
Helpers.host_versions_not_decreasing is called.
Also, a few line breaks.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>