CA-220610: Changed host.get_vms_which_prevent_evacutation to return a map with unique keys
The previous output was a map where duplicate keys were allowed. The xmlrpc
libraries used by the SDKs, which are autogenerated from the API, cope with this
situation in different ways, namely the C library handles it ok, Python and .NET
keep only one of the offending keys, while Java throws explicitly an exception.
The call now has been changed to return a map with unique keys. An additional
call returning the old output has been added for the benefit of the CLI.

Signed-off-by: Konstantina Chremmou <konstantina.chremmou@citrix.com>