CA-222760: Yangtze Repacked inspur hotfix for TLS Cream: The uploaded patch file is invalid
This issue only happend in Inspur hotfix.
Inspur signs their hotfixes with their own key.
So, Inspur's expectation is configured by a line starting "hotfix-fingerprint =" in xapi.conf which overrides xapi's built-in default.
The way to fix this is to copy the cipher defaults into xapi rather than into the config file.

Signed-off-by: Liang Dai <liang.dai1@citrix.com>