create VM gust metrics at first update
This commit moves the check for `last_start_time` on top of the
others to make sure that stale guest metrics, if present, are deleted
immediately. In addition it adds the logic to generate the VM guest
metrics if they are not yet present. In practice this means that the
guest metrics are generated during the VM start, and are kept updated
trhoughout the update VM process.

Signed-off-by: Marcello Seri <marcello.seri@citrix.com>