CA-224730 back off exponentially after failed read from plugin
This commit introduces a skip count for plugins. After a read error, the
skip count is incremented and the plugin is not read for that number of
read cycles. When reading succeeds, the skip is reset to 0. Currently,
at most 256 cycles are skipped and the skip count increases
exponentially from 1 with every read error.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>