CA-224967: Don't remove proxy state from xenstore when OVS rules are removed
When a proxied VM is rebooted, the pvsproxy daemon is left configured, but the
OVS rules are removed and put back for the new VIF. This triggers and event
from xenopsd to xapi to update PVS_proxy.currently_attached to false. Xapi also
happens to remove the pvsproxy state from xenstore, even though the daemon is
still configured. This causes the OVS rules to not be reinstated when the VM
comes up again.

The solution is to simply leave the pvsproxy state alone.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>