CA-224975: PVS-cache: Handle cache VDIs on non-persistent SR
PVS-cache supports a new kind of SR that is based on a RAM disk. A consequence
is that the contents of the SR disappear when the host is rebooted, while
leaving the VDI records behind in the database, and PVS_cache_storage.VDI
fields may be invalid.

To handle this, we now check whether the PVS_cache_storage.VDI that is needed
when a PVS-cached VM is started still actually exists. If not, it will be
recreated. Furthermore, in PVS_cache_storage.destroy, we ignore errors about
the VDI not existing anymore.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>