CP-18612: Introduce multiple flow tables for pvs vswitch rules (#293)
Introduce separate tables for IP addresses (table 0), UDP port numbers
(table 101) and actions (table 102). This reduces the number of flows
added, to vswitch, by 10 folds.

* Use cookie to tag per client pvs vswitch rule

pvs<mac> interface is created per client and ifindex of this interface
will be distinct, for each of the clients, in a host. Use ifindex as a
cookie to tag all rules created for a client. It make easier to remove
per client vswitch rules.

* Store ofport and ifindex, of a vif/tap, in xenstore

While adding vif or tap interface, ifindex and ofport of this interface will
be stored under /xapi/<bla>/private/vif/x/y/. This information can be used
by setup-*-rules to create rules based on ifindex/ofport.

* Use VM's interface index as cookie

Use vm's interface index as cookie. tap/vif interface have different
ifindex and it is easy to delete rules if rules are taged with this cookie.

Signed-off-by: Anoob Soman <anoob.soman@citrix.com>