CA-223506: setup-pvs-proxy-rules: handle localhost migration
Localhost migration failed, because the vif-real and setup-pvs-proxy-rules
scripts use xenstore keys to pass information between each other and xenopsd,
but the keys were in the VM-UUID-based private tree. During a localhost
migration, the sender writes keys that are soon after removed by the receiver.

We solve this by moving the xenstore keys from the private path to the hotplug
path, which is indexed by domid.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>