CA-224335 re-start agetty only when IP addr changes; fix signaling this
This commit addresses agetty being restarted so frequently that systemd
would block it. We now limit restarts to situation where we know that
the IP address has changed. This should bring down the frequency.

This commit fixes signaling of the IP address change: the change
previously was part of an exception handler and was almost never
executed. It is now executed unconditionally.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>