CA-226177: Fix premature termination of cli
When the CLI makes a TLS connection to a remote host as part of a
command, there was code to monitor the stunnel process and exit
early if the stunnel process died. There was a race, therefore,
when a TLS connection terminiated correctly but the command hadn't
finished processing on the control socket, for example, when
uploading a VDI via a CLI on a slave. This commit removes the active
monitoring of stunnel processes.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>