CA-227721: init-script: flush to log on xapi start
The start function in the init-script for xapi was writing
"Starting xapi: " with no newline on success, so the text
did not reach the logging system until the next stop; this
caused wrong and confusing logs.

Now the function emits a newline immediately before it
replaces itself with a xapi process.

Signed-off-by: Thomas Sanders <thomas.sanders@citrix.com>