CA-225719: automonously enable/disable redo_log on PBD plug/unplug
Nothing special when we do this on a slave PBD.

But in case when we unplug the relevant PBD on the master, this would
pause the redo_log service for the time being. But situation like this (master
PBD unplugged) usually won't last long and the service will recover as soon as
the PBD is plug back. If we consider the implicit redo_log enabling is a best
effort thing, this is a comparatively clean/simple solution.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>