CA-225070: shutdown HA daemon explicitly in case of host shutdown
HA shutdown is not part of the normal toolstack shutdown/restart process, as HA
daemon is designed to outlive xapi.

This is not a very clean solution (with reference to absolute paths of the
cluster_stack). A sligtly better way is to make use of some HA daemon
shutdown (temporarily) function, which isn't exposed at the moment.

Signed-off-by: Zheng Li <zheng.li3@citrix.com>