CP-19645: qemu-dm-wrapper: run QEMU if requested
If the VM parameter xenstore-data:vm-data/up-qemu is "1" (or "2" and
the guest is being restored), run QEMU instead of qemu-trad.

Since the command line arguments don't match between qemu-trad and
QEMU, exec the wrapper script provided by the qemu package to
transform the arguments.

If xenstore-data:vm-data/up-qemu is "0" or not set (the default), then
qemu-trad is used as before.