Remove FORK_USER variable from .travis.yml
The travis build fails because we do not have a fork of the
ocaml-ci-scripts repository, which .travis-opam.sh tries to use if
FORK_USER is specified.

We do have a fork of ocaml-ci-scripts with a different name, called
"ocaml-travisci-skeleton"
(https://github.com/xapi-project/ocaml-travisci-skeleton), but this
seems to be outdated.