CA-228573: Require a content-length on pool database restore HTTP handler
This fixes an issue where the CLI was hanging when restoring the database.
The cause was that the CLI is no longer terminating the stunnel after
sending the file as this was leading to data loss. The handler was
therefore waiting for more data forever. This patch requires the
content-length header and terminates the connection after receiving
that exact number of bytes.

Signed-off-by: Jon Ludlam <jonathan.ludlam@citrix.com>