CA-229176 fix PVS Proxy status cache inconsistency: invalidate during attach
This commits fix the problem that the xapi database would not contain
the correct status of a PVS proxy.

The problem is rooted in a caching layer: the status as reported via RRD
from PVS proxies is read into a cache. When the cache is updated, this
update is propagated to the Xapi database. However, the Xapi database is
updated only for attached proxies. This leads to an inconsistency: when
the PVS Proxy reached the state of "caching" before it was attached, the
Xapi entry would never be updated, staying in status "initialised". The
problem is solved by invalidating the cache entry once the PVS proxy is
attached. Now the current state is read into the cache and propagated to
the Xapi database.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>